Demote auto update message to Debug level
The message was spamming the console in result of what is a very intentional change of run parameters. If an user decides they didn't want auto update to run via launch parameters, they likely had a good reason and are not needed to be reminded so frequently about that decision.